DEV: Move shared drag-and-drop helpers under `lib/-internals`
The consumer guard, the drag vocabulary, the external payload helpers and the test reset lived in the `dragAndDrop` service module only because the modifiers could not import one another without closing a cycle. They now live in `lib/-internals/drag-and-drop/`, the marker `lib/blocks/-internals/` already uses for code that is not a public import surface, and the service module keeps the service alone. Behaviour-preserving: symbols moved verbatim and importers rewired.
